Add AnswerRepository & QuestionRepository

+++ src/main/java/survey/backend/repository/SurveyRepository.java
@@ -3,5 +3,5 @@ package survey.backend.repository;
 import org.springframework.data.repository.CrudRepository;
 import survey.backend.entities.Survey;
 
-public interface SurveyRepository extends CrudRepository<Survey, Integer> {
+public interface SurveyRepository extends CrudRepository<Survey, Long> {
 }
